Final use case fixes
