+++ catowseriOS/Domain/Sources/CottonUseCases/UseCases/AddTabUseCase.swift
@@ -8,6 +8,7 @@
 
 import AutoMockable
 import CoreBrowser
+import CottonTabs
 import BaseUseCaseKit
 
 /// Add tab use case.
@@ -17,4 +18,25 @@ public protocol AddTabUseCase: CoreUseCase, AutoMockable, Sendable {
     ///
     /// - Parameter tab: A tab.
     func execute(input: CoreBrowser.Tab) async throws
+}
+
+public final class AddTabUseCaseImpl: AddTabUseCase {
+    private let tabsDataService: any TabsDataServiceProtocol
+
+    public init(_ tabsDataService: any TabsDataServiceProtocol) {
+        self.tabsDataService = tabsDataService
+    }
+
+    public func execute(input tab: CoreBrowser.Tab) async throws {
+        let serviceData = await tabsDataService.sendCommand(.addTab(tab), nil)
+        guard case let .finished(result) = serviceData.tabAdded else {
+            throw AppError.commandNotFinishedYet
+        }
+        switch result {
+        case .failure(let error):
+            throw error
+        case .success:
+            return
+        }
+    }
 }
