+++ catowseriOS/Domain/Sources/CottonUseCases/UseCases/SelectTabUseCase.swift
@@ -8,6 +8,7 @@
 
 import AutoMockable
 import CoreBrowser
+import CottonTabs
 import BaseUseCaseKit
 
 /// Select tab use case.
@@ -16,4 +17,25 @@ public protocol SelectTabUseCase: CoreUseCase, AutoMockable, Sendable {
     ///
     /// - Parameter tab: A tab to select
     func execute(input: CoreBrowser.Tab) async throws
+}
+
+public final class SelectTabUseCaseImpl: SelectTabUseCase {
+    private let tabsDataService: any TabsDataServiceProtocol
+
+    public init(_ tabsDataService: any TabsDataServiceProtocol) {
+        self.tabsDataService = tabsDataService
+    }
+
+    public func execute(input tab: CoreBrowser.Tab) async throws {
+        let serviceData = await tabsDataService.sendCommand(.selectTab(tab), nil)
+        guard case let .finished(result) = serviceData.tabSelected else {
+            throw AppError.commandNotFinishedYet
+        }
+        switch result {
+        case .failure(let error):
+            throw AppError.tabsServiceError(error)
+        case .success:
+            return
+        }
+    }
 }
